Add docs and tests
