+++ crates/ruff/src/rules/pylint/mod.rs
@@ -31,6 +31,7 @@ mod tests {
         Rule::RepeatedIsinstanceCalls,
         Path::new("repeated_isinstance_calls.py")
     )]
+    #[test_case(Rule::ComparisonWithItself, Path::new("comparison_with_itself.py"))]
     #[test_case(Rule::ManualFromImport, Path::new("import_aliasing.py"))]
     #[test_case(Rule::SysExitAlias, Path::new("sys_exit_alias_0.py"))]
     #[test_case(Rule::SysExitAlias, Path::new("sys_exit_alias_1.py"))]
