+++ ruff.schema.json
@@ -2108,6 +2108,8 @@
         "PLR",
         "PLR0",
         "PLR01",
+        "PLR012",
+        "PLR0124",
         "PLR013",
         "PLR0133",
         "PLR02",
